Restore model data word-pointer types

+++ src/main.exe/GetWeaponData.c
@@ -90,9 +90,9 @@ extern WeaponModelType WeaponModel[41];
 extern char D_800117EC[]; /* "%s%s.TMD" */
 extern char D_800117F8[]; /* "K:\\WORK\\CDIMAGE\\HUMAN\\WEAPON\\" */
 
-extern u32 *FileRead(u8 *filename);
+extern u_long *FileRead(u8 *filename);
 extern int sprintf(char *buf, char *fmt, ...);
-extern OrnamentType *LoadOrnament(u32 *adr);
+extern OrnamentType *LoadOrnament(u_long *adr);
 
 static inline void FindWeaponId(Humanoid *human, s16 wid, s16 wpid)
 {
